Finish app

+++ lib/screens/favorites_screen/favorites_view_model.dart
@@ -1,3 +1,4 @@
+import 'package:ar_cademy/core/manager/app_provider.dart';
 import 'package:flutter/material.dart';
 
 import '../../core/web_services/firebase_utils.dart';
@@ -16,38 +17,43 @@ class FavoritesViewModel extends ChangeNotifier {
   bool get hasError => _hasError;
 
   Future<void> getFavorites() async {
-    _isLoading = true;
-    _hasError = false;
-    notifyListeners();
+    if (AppProvider.user == null && AppProvider.userId == "null") {
+      _isLoading = true;
+      _hasError = false;
+      notifyListeners();
 
-    final snapshot = await FirebaseUtils.getUserData();
-    snapshot.listen((querySnapshot) {
-      try {
-        Map<String, dynamic> data =
-            querySnapshot.data()! as Map<String, dynamic>;
-        List<dynamic> favoriteItemIds = data["favoriteItemIds"];
+      final snapshot = await FirebaseUtils.getUserData();
+      snapshot.listen((querySnapshot) {
+        try {
+          Map<String, dynamic> data =
+          querySnapshot.data()! as Map<String, dynamic>;
+          List<dynamic> favoriteItemIds = data["favoriteItemIds"];
 
-        _myFavorites.clear();
-        _addToMyFavorites(favoriteItemIds);
+          _myFavorites.clear();
+          _addToMyFavorites(favoriteItemIds);
 
-        _isLoading = false;
-        notifyListeners();
-      } catch (e) {
-        debugPrint("$e");
-        _isLoading = false;
-        _hasError = true;
-        notifyListeners();
-      }
-    });
+          _isLoading = false;
+          notifyListeners();
+        } catch (e) {
+          debugPrint("$e");
+          _isLoading = false;
+          _hasError = true;
+          notifyListeners();
+        }
+      });
+    }
+
   }
 
   void _addToMyFavorites(List<dynamic> favoriteItemIds) {
-    for (final itemId in favoriteItemIds) {
-      for (final systemOrgans in SystemsData.systemsOrgans.values) {
-        for (final organ in systemOrgans) {
-          if (organ.id == itemId) {
-            _myFavorites.add(organ..isFavorite = true);
-            break;
+    if (AppProvider.user == null && AppProvider.userId == "null") {
+      for (final itemId in favoriteItemIds) {
+        for (final systemOrgans in SystemsData.systemsOrgans.values) {
+          for (final organ in systemOrgans) {
+            if (organ.id == itemId) {
+              _myFavorites.add(organ..isFavorite = true);
+              break;
+            }
           }
         }
       }
@@ -55,12 +61,16 @@ class FavoritesViewModel extends ChangeNotifier {
   }
 
   void addToFavorites(OrganModel organ) {
-    FirebaseUtils.addToFavorites(itemId: organ.id);
-    notifyListeners();
+    if (AppProvider.user == null && AppProvider.userId == "null") {
+      FirebaseUtils.addToFavorites(itemId: organ.id);
+      notifyListeners();
+    }
   }
 
   void deleteFromFavorites(OrganModel organ) {
-    FirebaseUtils.deleteFromFavorites(itemId: organ.id);
-    notifyListeners();
+    if (AppProvider.user == null && AppProvider.userId == "null") {
+      FirebaseUtils.deleteFromFavorites(itemId: organ.id);
+      notifyListeners();
+    }
   }
 }
